+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [next]
 
+### Added
+
+- Add RubyMine `com.jetbrains.intellij.rubymine:rubymine` Maven coordinates to the `IntelliJPlatformType`
+
 ### Changed
 
 - Update minimal supported Gradle version to `8.6`
@@ -11,6 +15,7 @@
 - Add a check for target file existence before creating and writing to `JarOutputStream`. JetBrains/intellij-platform-gradle-plugin#1988
 - Fix the incorrect rsync exclusion pattern in `DmgExtractorValueSource`
 - Fix macOS missing nio-fs.jar on bootclasspath when runIde. JetBrains/intellij-platform-gradle-plugin#1994
+- Pass sandbox properties path with invariant separators.
 
 ## [2.7.0] - 2025-07-26